jvss: moving old code to separate folder
