+++ protocols/jvss_old/simulation.go
@@ -1,4 +1,4 @@
-package jvss
+package jvss_old
 
 import (
 	"github.com/BurntSushi/toml"
